[Silver IV] Title: 스케이트 연습, Time: 292 ms, Memory: 3976 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/28324. 스케이트 연습/스케이트 연습.cc
@@ -10,8 +10,8 @@ using namespace std;
 
 int n;
 vector<int> arr(n);
-int r_sum = 0;
-int num = 1;
+long long r_sum = 0;
+int num = 1; // 높일 수 있는 max
 
 void solution() {
 	for (int i = n - 1; i >= 0; i--) {
